refactor(webhook): extract profile admission policy
